[TEST] Use new flag 0x2001 for floating window
This is to comply with test xhfw mods. Lollipop uses the old flag (0x2000) as FLAG_ACTIVITY_RETAIN_IN_RECENTS - to keep list of documents or open pages in recent tasks when the user closes the app. It is default launch flag for the new Chrome, for example. Getting the new unused flag - 0x2001.